+++ node/service/src/client.rs
@@ -1,1 +1,399 @@
+// Copyright 2019-2021 Bholdus Inc.
+// This file is part of Bholdus.
 
+// Bholdus is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Bholdus is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Bholdus.  If not, see <http://www.gnu.org/licenses/>.
+use crate::{AccountId, Balance, Block, BlockNumber, Hash, Header, Index};
+use sc_client_api::{Backend as BackendT, BlockchainEvents, KeyIterator, PrefixedStorageKey};
+use sp_api::{CallApiAt, NumberFor, ProvideRuntimeApi};
+use sp_blockchain::HeaderBackend;
+use sp_consensus::BlockStatus;
+use sp_consensus_aura::sr25519::AuthorityId as AuraId;
+use sp_runtime::{
+    generic::{BlockId, SignedBlock},
+    traits::{BlakeTwo256, Block as BlockT},
+    Justifications,
+};
+use sp_storage::{ChildInfo, StorageData, StorageKey};
+use std::sync::Arc;
+
+/// A set of APIs that bholdus-like runtimes must implement.
+///
+/// This trait has no methods or associated type. It is a concise marker for all the trait bounds
+/// that it contains.
+pub trait RuntimeApiCollection:
+    sp_api::ApiExt<Block>
+    + sp_api::Metadata<Block>
+    + sp_authority_discovery::AuthorityDiscoveryApi<Block>
+    + sp_block_builder::BlockBuilder<Block>
+    + sp_consensus_aura::AuraApi<Block, AuraId>
+    + sp_finality_grandpa::GrandpaApi<Block>
+    + beefy_primitives::BeefyApi<Block>
+    + sp_offchain::OffchainWorkerApi<Block>
+    + sp_session::SessionKeys<Block>
+    + sp_transaction_pool::runtime_api::TaggedTransactionQueue<Block>
+    + frame_system_rpc_runtime_api::AccountNonceApi<Block, AccountId, Index>
+    + pallet_transaction_payment_rpc_runtime_api::TransactionPaymentApi<Block, Balance>
+    + pallet_contracts_rpc::ContractsRuntimeApi<Block, AccountId, Balance, BlockNumber, Hash>
+    + pallet_mmr_rpc::MmrRuntimeApi<Block, <Block as sp_runtime::traits::Block>::Hash>
+    + fp_rpc::EthereumRuntimeRPCApi<Block>
+where
+    <Self as sp_api::ApiExt<Block>>::StateBackend: sp_api::StateBackend<BlakeTwo256>,
+{
+}
+
+impl<Api> RuntimeApiCollection for Api
+where
+    Api: sp_api::ApiExt<Block>
+        + sp_api::Metadata<Block>
+        + sp_authority_discovery::AuthorityDiscoveryApi<Block>
+        + sp_block_builder::BlockBuilder<Block>
+        + sp_consensus_aura::AuraApi<Block, AuraId>
+        + sp_finality_grandpa::GrandpaApi<Block>
+        + beefy_primitives::BeefyApi<Block>
+        + sp_offchain::OffchainWorkerApi<Block>
+        + sp_session::SessionKeys<Block>
+        + sp_transaction_pool::runtime_api::TaggedTransactionQueue<Block>
+        + frame_system_rpc_runtime_api::AccountNonceApi<Block, AccountId, Index>
+        + pallet_transaction_payment_rpc_runtime_api::TransactionPaymentApi<Block, Balance>
+        + pallet_contracts_rpc::ContractsRuntimeApi<Block, AccountId, Balance, BlockNumber, Hash>
+        + pallet_mmr_rpc::MmrRuntimeApi<Block, <Block as sp_runtime::traits::Block>::Hash>
+        + fp_rpc::EthereumRuntimeRPCApi<Block>,
+    <Self as sp_api::ApiExt<Block>>::StateBackend: sp_api::StateBackend<BlakeTwo256>,
+{
+}
+
+/// Config that abstracts over all available client implementations.
+///
+/// For a concrete type there exists [`Client`].
+pub trait AbstractClient<Block, Backend>:
+    BlockchainEvents<Block>
+    + Sized
+    + Send
+    + Sync
+    + ProvideRuntimeApi<Block>
+    + HeaderBackend<Block>
+    + CallApiAt<Block, StateBackend = Backend::State>
+where
+    Block: BlockT,
+    Backend: BackendT<Block>,
+    Backend::State: sp_api::StateBackend<BlakeTwo256>,
+    Self::Api: RuntimeApiCollection<StateBackend = Backend::State>,
+{
+}
+
+impl<Block, Backend, Client> AbstractClient<Block, Backend> for Client
+where
+    Block: BlockT,
+    Backend: BackendT<Block>,
+    Backend::State: sp_api::StateBackend<BlakeTwo256>,
+    Client: BlockchainEvents<Block>
+        + ProvideRuntimeApi<Block>
+        + HeaderBackend<Block>
+        + Sized
+        + Send
+        + Sync
+        + CallApiAt<Block, StateBackend = Backend::State>,
+    Client::Api: RuntimeApiCollection<StateBackend = Backend::State>,
+{
+}
+
+/// Execute something with the client instance.
+///
+/// As there exist multiple chains inside Bholdus, like Ulas, Cygnus,
+/// Phoenix etc, there can exist different kinds of client types. As these
+/// client types differ in the generics that are being used, we can not easily
+/// return them from a function. For returning them from a function there exists
+/// [`Client`]. However, the problem on how to use this client instance still
+/// exists. This trait "solves" it in a dirty way. It requires a type to
+/// implement this trait and than the [`execute_with_client`](ExecuteWithClient:
+/// :execute_with_client) function can be called with any possible client
+/// instance.
+///
+/// In a perfect world, we could make a closure work in this way.
+pub trait ExecuteWithClient {
+    /// The return type when calling this instance.
+    type Output;
+
+    /// Execute whatever should be executed with the given client instance.
+    fn execute_with_client<Client, Api, Backend>(self, client: Arc<Client>) -> Self::Output
+    where
+        <Api as sp_api::ApiExt<Block>>::StateBackend: sp_api::StateBackend<BlakeTwo256>,
+        Backend: sc_client_api::Backend<Block>,
+        Backend::State: sp_api::StateBackend<BlakeTwo256>,
+        Api: crate::RuntimeApiCollection<StateBackend = Backend::State>,
+        Client: AbstractClient<Block, Backend, Api = Api> + 'static;
+}
+
+/// A handle to a Moonbeam client instance.
+///
+/// The Moonbeam service supports multiple different runtimes (Moonbase, Moonbeam
+/// itself, etc). As each runtime has a specialized client, we need to hide them
+/// behind a trait. This is this trait.
+///
+/// When wanting to work with the inner client, you need to use `execute_with`.
+pub trait ClientHandle {
+    /// Execute the given something with the client.
+    fn execute_with<T: ExecuteWithClient>(&self, t: T) -> T::Output;
+}
+
+/// A client instance of Moonbeam.
+#[derive(Clone)]
+pub enum Client {
+    #[cfg(feature = "with-ulas-runtime")]
+    Ulas(Arc<crate::FullClient<ulas_runtime::RuntimeApi, crate::UlasExecutor>>),
+    #[cfg(feature = "with-cygnus-runtime")]
+    Cygnus(Arc<crate::FullClient<cygnus_runtime::RuntimeApi, crate::CygnusExecutor>>),
+    #[cfg(feature = "with-phoenix-runtime")]
+    Phoenix(Arc<crate::FullClient<phoenix_runtime::RuntimeApi, crate::PhoenixExecutor>>),
+}
+
+#[cfg(feature = "with-ulas-runtime")]
+impl From<Arc<crate::FullClient<ulas_runtime::RuntimeApi, crate::UlasExecutor>>> for Client {
+    fn from(client: Arc<crate::FullClient<ulas_runtime::RuntimeApi, crate::UlasExecutor>>) -> Self {
+        Self::Ulas(client)
+    }
+}
+
+#[cfg(feature = "with-cygnus-runtime")]
+impl From<Arc<crate::FullClient<cygnus_runtime::RuntimeApi, crate::CygnusExecutor>>> for Client {
+    fn from(
+        client: Arc<crate::FullClient<cygnus_runtime::RuntimeApi, crate::CygnusExecutor>>,
+    ) -> Self {
+        Self::Cygnus(client)
+    }
+}
+
+#[cfg(feature = "with-phoenix-runtime")]
+impl From<Arc<crate::FullClient<phoenix_runtime::RuntimeApi, crate::PhoenixExecutor>>> for Client {
+    fn from(
+        client: Arc<crate::FullClient<phoenix_runtime::RuntimeApi, crate::PhoenixExecutor>>,
+    ) -> Self {
+        Self::Phoenix(client)
+    }
+}
+
+impl ClientHandle for Client {
+    fn execute_with<T: ExecuteWithClient>(&self, t: T) -> T::Output {
+        match self {
+            #[cfg(feature = "with-ulas-runtime")]
+            Self::Ulas(client) => {
+                T::execute_with_client::<_, _, crate::FullBackend>(t, client.clone())
+            }
+            #[cfg(feature = "with-cygnus-runtime")]
+            Self::Cygnus(client) => {
+                T::execute_with_client::<_, _, crate::FullBackend>(t, client.clone())
+            }
+            #[cfg(feature = "with-phoenix-runtime")]
+            Self::Phoenix(client) => {
+                T::execute_with_client::<_, _, crate::FullBackend>(t, client.clone())
+            }
+        }
+    }
+}
+
+macro_rules! match_client {
+	($self:ident, $method:ident($($param:ident),*)) => {
+		match $self {
+			#[cfg(feature = "with-ulas-runtime")]
+			Self::Ulas(client) => client.$method($($param),*),
+			#[cfg(feature = "with-cygnus-runtime")]
+			Self::Cygnus(client) => client.$method($($param),*),
+			#[cfg(feature = "with-phoenix-runtime")]
+			Self::Phoenix(client) => client.$method($($param),*),
+		}
+	};
+}
+
+impl sc_client_api::UsageProvider<Block> for Client {
+    fn usage_info(&self) -> sc_client_api::ClientInfo<Block> {
+        match_client!(self, usage_info())
+    }
+}
+
+impl sc_client_api::BlockBackend<Block> for Client {
+    fn block_body(
+        &self,
+        id: &BlockId<Block>,
+    ) -> sp_blockchain::Result<Option<Vec<<Block as BlockT>::Extrinsic>>> {
+        match_client!(self, block_body(id))
+    }
+
+    fn block_indexed_body(
+        &self,
+        id: &BlockId<Block>,
+    ) -> sp_blockchain::Result<Option<Vec<Vec<u8>>>> {
+        match_client!(self, block_indexed_body(id))
+    }
+
+    fn block(&self, id: &BlockId<Block>) -> sp_blockchain::Result<Option<SignedBlock<Block>>> {
+        match_client!(self, block(id))
+    }
+
+    fn block_status(&self, id: &BlockId<Block>) -> sp_blockchain::Result<BlockStatus> {
+        match_client!(self, block_status(id))
+    }
+
+    fn justifications(&self, id: &BlockId<Block>) -> sp_blockchain::Result<Option<Justifications>> {
+        match_client!(self, justifications(id))
+    }
+
+    fn block_hash(
+        &self,
+        number: NumberFor<Block>,
+    ) -> sp_blockchain::Result<Option<<Block as BlockT>::Hash>> {
+        match_client!(self, block_hash(number))
+    }
+
+    fn indexed_transaction(
+        &self,
+        hash: &<Block as BlockT>::Hash,
+    ) -> sp_blockchain::Result<Option<Vec<u8>>> {
+        match_client!(self, indexed_transaction(hash))
+    }
+
+    fn has_indexed_transaction(
+        &self,
+        hash: &<Block as BlockT>::Hash,
+    ) -> sp_blockchain::Result<bool> {
+        match_client!(self, has_indexed_transaction(hash))
+    }
+}
+
+impl sc_client_api::backend::StorageProvider<Block, crate::FullBackend> for Client {
+    fn storage(
+        &self,
+        id: &BlockId<Block>,
+        key: &StorageKey,
+    ) -> sp_blockchain::Result<Option<StorageData>> {
+        match_client!(self, storage(id, key))
+    }
+
+    fn storage_keys(
+        &self,
+        id: &BlockId<Block>,
+        key_prefix: &StorageKey,
+    ) -> sp_blockchain::Result<Vec<StorageKey>> {
+        match_client!(self, storage_keys(id, key_prefix))
+    }
+
+    fn storage_hash(
+        &self,
+        id: &BlockId<Block>,
+        key: &StorageKey,
+    ) -> sp_blockchain::Result<Option<<Block as BlockT>::Hash>> {
+        match_client!(self, storage_hash(id, key))
+    }
+
+    fn storage_pairs(
+        &self,
+        id: &BlockId<Block>,
+        key_prefix: &StorageKey,
+    ) -> sp_blockchain::Result<Vec<(StorageKey, StorageData)>> {
+        match_client!(self, storage_pairs(id, key_prefix))
+    }
+
+    fn storage_keys_iter<'a>(
+        &self,
+        id: &BlockId<Block>,
+        prefix: Option<&'a StorageKey>,
+        start_key: Option<&StorageKey>,
+    ) -> sp_blockchain::Result<
+        KeyIterator<'a, <crate::FullBackend as sc_client_api::Backend<Block>>::State, Block>,
+    > {
+        match_client!(self, storage_keys_iter(id, prefix, start_key))
+    }
+
+    fn child_storage(
+        &self,
+        id: &BlockId<Block>,
+        child_info: &ChildInfo,
+        key: &StorageKey,
+    ) -> sp_blockchain::Result<Option<StorageData>> {
+        match_client!(self, child_storage(id, child_info, key))
+    }
+
+    fn child_storage_keys(
+        &self,
+        id: &BlockId<Block>,
+        child_info: &ChildInfo,
+        key_prefix: &StorageKey,
+    ) -> sp_blockchain::Result<Vec<StorageKey>> {
+        match_client!(self, child_storage_keys(id, child_info, key_prefix))
+    }
+
+    fn child_storage_keys_iter<'a>(
+        &self,
+        id: &BlockId<Block>,
+        child_info: ChildInfo,
+        prefix: Option<&'a StorageKey>,
+        start_key: Option<&StorageKey>,
+    ) -> sp_blockchain::Result<
+        KeyIterator<'a, <crate::FullBackend as sc_client_api::Backend<Block>>::State, Block>,
+    > {
+        match_client!(
+            self,
+            child_storage_keys_iter(id, child_info, prefix, start_key)
+        )
+    }
+
+    fn child_storage_hash(
+        &self,
+        id: &BlockId<Block>,
+        child_info: &ChildInfo,
+        key: &StorageKey,
+    ) -> sp_blockchain::Result<Option<<Block as BlockT>::Hash>> {
+        match_client!(self, child_storage_hash(id, child_info, key))
+    }
+
+    fn max_key_changes_range(
+        &self,
+        first: NumberFor<Block>,
+        last: BlockId<Block>,
+    ) -> sp_blockchain::Result<Option<(NumberFor<Block>, BlockId<Block>)>> {
+        match_client!(self, max_key_changes_range(first, last))
+    }
+
+    fn key_changes(
+        &self,
+        first: NumberFor<Block>,
+        last: BlockId<Block>,
+        storage_key: Option<&PrefixedStorageKey>,
+        key: &StorageKey,
+    ) -> sp_blockchain::Result<Vec<(NumberFor<Block>, u32)>> {
+        match_client!(self, key_changes(first, last, storage_key, key))
+    }
+}
+
+impl sp_blockchain::HeaderBackend<Block> for Client {
+    fn header(&self, id: BlockId<Block>) -> sp_blockchain::Result<Option<Header>> {
+        let id = &id;
+        match_client!(self, header(id))
+    }
+
+    fn info(&self) -> sp_blockchain::Info<Block> {
+        match_client!(self, info())
+    }
+
+    fn status(&self, id: BlockId<Block>) -> sp_blockchain::Result<sp_blockchain::BlockStatus> {
+        match_client!(self, status(id))
+    }
+
+    fn number(&self, hash: Hash) -> sp_blockchain::Result<Option<BlockNumber>> {
+        match_client!(self, number(hash))
+    }
+
+    fn hash(&self, number: BlockNumber) -> sp_blockchain::Result<Option<Hash>> {
+        match_client!(self, hash(number))
+    }
+}
