--- node/service/src/service/mod.rs
@@ -1,8 +1,0 @@
-#[cfg(feature = "with-ulas-runtime")]
-pub mod ulas;
-
-#[cfg(feature = "with-cygnus-runtime")]
-pub mod cygnus;
-
-#[cfg(feature = "with-phoenix-runtime")]
-pub mod phoenix;
